MB-6544 Fix segfault in compaction error cleanup
couchstore_close_db should not segfault when passed NULL, as may happen
in the cleanup and error handling of the compactor when it cannot open
the destination file.

also, avoid deleting files during cleanup that we didn't create.

Reviewed-on: http://review.couchbase.org/20596
Tested-by: Aaron Miller <apage43@ninjawhale.com>

+++ src/db_compact.c
@@ -66,7 +66,7 @@ cleanup:
         fclose(ctx.id_tmp);
     }
     couchstore_close_db(target);
-    if(errcode != COUCHSTORE_SUCCESS) {
+    if(errcode != COUCHSTORE_SUCCESS && target != NULL) {
         unlink(target_filename);
     }
     return errcode;
